Fix files, purchases index tests failing
